Split color picker into 2 files
